Failed attempt at bugfixing
Cannot for the life of me get the onClick navigates to work on this page, tried putting it on every element, tried it with home page, tried adding ids, tried changing from const to function, tried adding dedicated buttons. nothing makes clicking anything on this page redirect to the item specific page.